pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^3.0"
python = "^3.8"
drf-yasg = "^1.21.5"
djangorestframework = "^3.14.0"
django-crm = "^0.9.0"
djangorestframework-word-filter = "^0.2.0"
[tool.poetry.dev-dependencies]
